+++ HABApp/core/EventBus.py
@@ -16,13 +16,6 @@ _LOCK = threading.Lock()
 _EVENT_LISTENER: typing.Dict[str, typing.List[EventBusListener]] = {}
 
 
-# todo: make central HABApp topic (name) definition here
-
-
-def __get_listener_description(listener: EventBusListener) -> str:
-    return f'"{listener.topic}" (type {listener.event_filter})'
-
-
 @log_exception
 def post_event(topic: str, event):
     assert isinstance(topic, str), type(topic)
@@ -54,12 +47,12 @@ def add_listener(listener: EventBusListener):
 
         # don't add the same listener twice
         if listener in item_listeners:
-            _habapp_log.warning(f'Event listener for {__get_listener_description(listener)} has already been added!')
+            _habapp_log.warning(f'Event listener for {listener.desc()} has already been added!')
             return None
 
         # add listener
         item_listeners.append(listener)
-        _habapp_log.debug(f'Added event listener for {__get_listener_description(listener)}')
+        _habapp_log.debug(f'Added event listener for {listener.desc()}')
         return None
 
 
@@ -72,12 +65,12 @@ def remove_listener(listener: EventBusListener):
 
         # print warning if we try to remove it twice
         if listener not in item_listeners:
-            _habapp_log.warning(f'Event listener for {__get_listener_description(listener)} has already been removed!')
+            _habapp_log.warning(f'Event listener for {listener.desc()} has already been removed!')
             return None
 
         # remove listener
         item_listeners.remove(listener)
-        _habapp_log.debug(f'Removed event listener for {__get_listener_description(listener)}')
+        _habapp_log.debug(f'Removed event listener for {listener.desc()}')
 
 
 @log_exception
