+++ HABApp/core/const/topics.py
@@ -1,3 +1,5 @@
+import typing
+
 try:
     from typing import Final
 except ImportError:
@@ -8,3 +10,10 @@ ERRORS: Final = 'HABApp.Errors'
 
 RULES: Final = 'HABApp.Rules'
 PARAM: Final = 'HABApp.Parameters'
+
+
+ALL: typing.List[str] = [
+    WARNINGS, ERRORS,
+
+    RULES, PARAM,
+]
