+++ HABApp/core/items/__init__.py
@@ -1,3 +1,3 @@
-from .base_value import BaseValueItem
+from .base_valueitem import BaseValueItem
 from .item import Item
 from .item_color import ColorItem
